Compose: use named volumes for pipeline and storage service data
This is so that we can pre-create the docker volumes, e.g. to use NFS share instead of local dir.

The src folder is now the default storage service location. This is so that the archivematica-sampledata
folder is available. This only applies for dev environment; for production it is expected that NFS
shares will be used in place of local directories.