deployment: restructure project definitions in playbooks
The playbooks used to use 'repos' and 'images' lists to describe what we
wanted to build. This was inefficient and unwieldy, since the version/tag
was repeated and the link between image and repo were implicit.

With this change, we now use a 'projects' list, which includes both repo
and image information. This means that images now take the version from the
enclosing project, and there can be multiple images per project.