deployment: fix Ansible linting issues
This is mostly about ensuring that command tasks are guarded using a when
clause. To achieve this, the output of the git module was captured in a
register var, which is then used to iterate through the projects that
require updates because they have changed.

Unfortunately, having cleared ANSIBLE0012 by adding the "when" clause, we
then hit ANSIBLE0016, which wrongly claims that the tasks should be
handlers rather than tasks - which isn't possible because we need to use
the item and other variables.