Update Makefile
This pull request brings the existing Make targets in sync to what we're doing
directly via our playbooks (targets `clone` and `build` or `build-images`). It
adds two new targets `publish` and `help`. It uses Ansible tags to run certain
tasks of the playbooks selectively.

See the new bits in the README.md file for more details.

This closes #114.