compose: update nextcloud service to match requirements of latest image
Recent work on the nextcloud image means we need to update the compose
configuration to accommodate the changes. The nextcloud volumes are now
external and there is also a dependency on the redis service too.

I've also updated the documentation for external volumes, as this was a
bit out of date in light of recent changes.