compose: copy files into am-shib nginx instead of mount
For the nginx image in the am-shib config we were using volume mounts to
include custom files for the image. These are better done as part of a COPY
in the Dockerfile, as it is less messy and more portable to environments
other than Docker Compose.

All the files have been moved into the new rootfs folder, which is copied
in a single COPY command in the Dockerfile.

The CMD and EXPOSE statements have also been moved from the compose config
into the image.