changed shibboleth to use SSL proxy in front of HTTP service
The all-in-one nginx-with-shibboleth container has been replaced with a
new, separate, shib-sp-proxy container, which uses Apache2 instead of
nginx, because this is better supported than the nginx module.

We now have an SSL proxy (implemented with nginx), which fronts the SP
proxy (implemented with Apache2), which fronts the normal Archivematica
proxy (implemented with nginx), which fronts the actual Archivematica
services (implemented with gunicorn). A bit of a server soup but it works!