Add reset-processing-configs target to Makefiles
This deletes the builtin processing configs (automated and default)
so that MCP Server can reset to the original configs when restarted.

Called before restart-mcp-services, if RESET_PROCESSING_CONFIG env var
is set to true. It is defined as a target in its own right in case there
is a need to call it manually.