compose: add Shibboleth and SSL for nextcloud
We can now optionally enable Shibboleth for nextcloud as part of the am-shib
Compose configuration. This also causes nextcloud to be proxied by the
nginx-ssl container, so that it runs behind SSL on port 443 like the other
services.

NextCloud has its own SP implementation so doesn't use the shib-sp-proxy that
Archivematica does. The IdP has been updated to include NextCloud's SP in its
metadata providers.

Note that, with this commit, the am-shib compose config now depends on the
nextcloud config, since it overrides it. This effectively means that it is no
no longer possible to run Archivematica without NextCloud if Shibboleth is
enabled.

The NextCloud image has also been updated to build from the dev/shibboleth
branch and to use the build-apps make target so all apps are built. Once the
current NextCloud work has been approved and we have a semver release tag
then we can update the playbook to use that instead.