compose: make nextcloud part of main config
The provision of NextCloud is no longer optional. It's been moved into the
main qa/dev config, so is always deployed alongside Archivematica.

The reason for this change is twofold: 1) it reduces the number of compose
config files and therefore simplifies the configuration, 2) the am-shib
config now requires both Archivematica and NextCloud, so it no longer makes
sense to separate them.