compose: don't set NextCloud env vars in qa config
We don't want to set the NEXTCLOUD_HOST or NEXTCLOUD_PORT env vars in qa
config because NextCloud is able to detect the correct values itself. The
only time we need to use these is in the am-shib config, as there NextCloud
is behind the SSL proxy, so we need to override the auto-detected values.