compose: move minikine into separate config
The minikine service has been moved into the now "mock-aws" Docker Compose
config. Other services, such as dynalite, may also follow.

The MOCK_AWS make argument can be used to control whether or not the mock
services are started. Default value is 'true', but if set to 'false' then
it becomes possible to specify real AWS resources to use instead.

The RDSS_ADAPTER_KINESIS_ENDPOINT and related environment variables can be
used to specify an alternative AWS service to use instead of minikine. See the
comments in the updated .env file for details.