compose: add restart-all-services target to Makefile
This commit adds a new Makefile target that restarts all services and once
restarted, it restarts again the gearmand and its dependent services to ensure
they are started in the right order because sometimes gearmand doesn't start
with all the services at the same time when deploying with ENV=dev.

This new Makefile target was added to the all target because it is needed to
restart the services at the end of the `make all` command, specially when using
Shibboleth.